pyproject.toml
[tool.poetry]
name = "al-arr-sync"
version = "0.1.0"
description = ""
authors = ["strNophix <nvdpoel01@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^0.20.0"
requests = "^2.28.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = {version = "^22.6.0", allow-prereleases = true}
pre-commit = "^2.20.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
